chore: rename feature to module
