+++ src/storeonwheels.client/src/app/libs/broadcast-module/components/geo-permission-instruction/geo-permission-instruction.component.ts
@@ -2,7 +2,7 @@ import { CommonModule } from "@angular/common";
 import { Component, Input } from "@angular/core";
 import { MatButtonModule } from "@angular/material/button";
 import { MatInputModule } from "@angular/material/input";
-import { GeoPermission } from "~/app/libs/broadcast-feature/services";
+import { GeoPermission } from "~/app/libs/broadcast-module/services";
 
 @Component({
   selector: "app-geo-permission-instruction",
