+++ src/storeonwheels.client/src/app/libs/map-module/services/timed-cache/timed-cache.service.spec.ts
@@ -1,7 +1,7 @@
 import { TestBed } from "@angular/core/testing";
 
 import { TickerCache } from "./ticker-cache.service";
-import { SleepService } from "~/app/libs/shared/services";
+import { SleepService } from "~/app/libs/shared-module";
 
 describe("TimeoutCacheService", () => {
   let cache: TickerCache<string, number>;
