+++ src/storeonwheels.client/src/app/pages/broadcast-page/broadcast-page.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import BroadcastPageComponent from "./broadcast-page.component";
 import axios from "axios";
-import { HUB_CONNECTION, hubConnection } from "~/app/libs/map-feature/services";
+import { HUB_CONNECTION, hubConnection } from "~/app/libs/map-module";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 
 describe("BroadcastPageComponent", () => {
