+++ src/storeonwheels.client/src/app/pages/healthcheck-page/healthcheck-page.component.ts
@@ -1,10 +1,7 @@
 import { Component, OnInit } from "@angular/core";
 import { HubConnectionState } from "@microsoft/signalr";
 import axios from "axios";
-import {
-  MessageHubService,
-  hubConnection,
-} from "~/app/libs/map-feature/services";
+import { MessageHubService, hubConnection } from "~/app/libs/map-module";
 
 @Component({
   selector: "app-healthcheck-page",
