+++ src/storeonwheels.client/src/app/pages/map-page/map-page.component.spec.ts
@@ -4,11 +4,11 @@ import {
   HUB_CONNECTION,
   MessageHubService,
   hubConnection,
-} from "~/app/libs/map-feature/services";
-import { GeoInfo, Vendor } from "~/app/libs/shared/models";
+} from "~/app/libs/map-module";
+import { GeoInfo, Vendor } from "~/app/libs/shared-module";
 import { Observable, from } from "rxjs";
 import { Mock } from "ts-mocks";
-import { VendorTableComponent } from "~/app/libs/map-feature/components";
+import { VendorTableComponent } from "~/app/libs/map-module";
 
 describe("MapPageComponent", () => {
   let component: MapPageComponent;
